+++ tests/DynamoDbTestCase.php
@@ -3,26 +3,20 @@
 namespace BaoPham\DynamoDb\Tests;
 
 use BaoPham\DynamoDb\DynamoDbServiceProvider;
+use Orchestra\Testbench\TestCase;
 
 /**
- * Class ModelTest
+ * Class DynamoDbTestCase
  *
  * @package BaoPham\DynamoDb\Tests
  */
-abstract class ModelTest extends TestCase
+abstract class DynamoDbTestCase extends TestCase
 {
-    /**
-     * @var \BaoPham\DynamoDb\DynamoDbModel
-     */
-    protected $testModel;
-
     public function setUp()
     {
         parent::setUp();
 
         $this->setUpDatabase();
-
-        $this->testModel = $this->getTestModel();
     }
 
     protected function getPackageProviders($app)
@@ -49,15 +43,8 @@ abstract class ModelTest extends TestCase
         ]);
     }
 
-    abstract protected function getTestModel();
-
     protected function setUpDatabase()
     {
         copy(dirname(__FILE__) . '/../dynamodb_local_init.db', dirname(__FILE__) . '/../dynamodb_local_test.db');
-    }
-
-    protected function getClient()
-    {
-        return $this->testModel->getClient();
     }
 }
