+++ tests/Parsers/UpdateExpressionTest.php
@@ -4,7 +4,7 @@ namespace BaoPham\DynamoDb\Tests\Parsers;
 
 use BaoPham\DynamoDb\Parsers\ExpressionAttributeNames;
 use BaoPham\DynamoDb\Parsers\UpdateExpression;
-use BaoPham\DynamoDb\Tests\TestCase;
+use PHPUnit\Framework\TestCase;
 
 class UpdateExpressionTest extends TestCase
 {
