--- tests/TestCase.php
@@ -1,12 +1,0 @@
-<?php
-
-namespace BaoPham\DynamoDb\Tests;
-
-/**
- * Class TestCase
- *
- * @package BaoPham\DynamoDb\Tests
- */
-abstract class TestCase extends \Orchestra\Testbench\TestCase
-{
-}
